--- src/main/java/com/example/utility/Controller.java
@@ -1,18 +1,0 @@
-package com.example.utility;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class Controller {
-    @GetMapping("/hi")
-    public String sayHi(){
-        return "Hello Hitesh";
-    }
-
-    @GetMapping("/hello")
-    public String sayHello(){
-        return "Hi Hitesh";
-    }
-
-}
