+++ src/main/resources/application.yml
@@ -1,0 +1,8 @@
+server:
+  servlet:
+    context-path: /my-app
+  port: 8082
+
+spring:
+  application:
+    name: utility
